Module Tjr_btree

Export the main functionality of this library.

Usage: open Tjr_btree;; open Tjr_btree.Btree_intf;;

Main interfaces

module Btree_intf : sig ... end

A collection of the main interfaces. This should be read first.

Disk_to_store

include Tjr_btree__.Disk_to_store

val disk_to_store : monad_ops:'t Tjr_monad.monad_ops -> disk_ops:('r, 't, 'dnode, 'blk) Btree_intf.disk_ops -> ('r, 'dnode, 't) Isa_btree.store_ops

Store to pre-btree

Pre-btree to map

include Tjr_btree__.Pre_btree_to_map

val pre_btree_to_map : monad_ops:'t Tjr_monad.monad_ops -> pre_btree_ops:('k, 'v, 'r, 't, 'leaf, 'node, 'leaf_stream) Isa_btree.pre_btree_ops -> root_ops:('r, 't) Btree_intf.btree_root_ops -> ('k, 'v, 'r, 'leaf_stream, 't) Btree_intf.Map_ops_with_ls.map_ops_with_ls

Main functionality: disk to map

module Make_1 : sig ... end

Main functionality: disk to map

include Make_1

module type S = sig ... end

This interface should be equal to Isa_btree_intf.S. Duplicated here so that ocamldoc can exhibit the definition.

module type T = sig ... end

module Make = Make_1.Make

The basic Make functor, with most things parameterizable. See also the more refined Tjr_btree_examples.Make_example functor.

Alternative version, using binprot, with simpler intf

module Make_2 : sig ... end

This is like make_1, but we assume binprot marshalling for k,v,r. This means we can simplify the interface considerably.

Alternative version, using binprot, with objects and no type generation

module Make_3 : sig ... end

This is like make_2, but we try to use objects only, no functors and no type generation.